requirements.txt
psutil>=5.9,<6.1
flask>=2.3.0,<4.0.0
flask-cors>=4.0.0,<5.0.0
requests>=2.31.0,<3.0.0
azure-storage-blob
python-dotenv
